% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_discount_factor.R
\name{apply_discount_factor}
\alias{apply_discount_factor}
\title{Apply Framework Discount Factor}
\usage{
apply_discount_factor(
  effect_size,
  discount_factor = 0.75,
  apply_discount = TRUE
)
}
\arguments{
\item{effect_size}{Input effect size (any metric)}

\item{discount_factor}{Discount factor (default: 0.75 per framework)}

\item{apply_discount}{Whether to apply discount (default: TRUE)}
}
\value{
Discounted effect size
}
\description{
Apply Framework Discount Factor
}
