% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon_signed_rank_power.R
\name{calculate_are_factor}
\alias{calculate_are_factor}
\title{Calculate Asymptotic Relative Efficiency Factor}
\usage{
calculate_are_factor()
}
\value{
ARE factor for Wilcoxon vs t-test
}
\description{
Calculate Asymptotic Relative Efficiency Factor
}
