% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_lagged_panel_power.R
\name{calculate_cross_lagged_effective_n}
\alias{calculate_cross_lagged_effective_n}
\title{Calculate Effective Sample Size for Cross-Lagged Analysis}
\usage{
calculate_cross_lagged_effective_n(n, stability_coefficient, n_waves)
}
\arguments{
\item{n}{Nominal sample size}

\item{stability_coefficient}{Stability coefficient}

\item{n_waves}{Number of waves}
}
\value{
Effective sample size
}
\description{
Calculate Effective Sample Size for Cross-Lagged Analysis
}
