% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_lagged_panel_power.R
\name{calculate_cross_lagged_required_n}
\alias{calculate_cross_lagged_required_n}
\title{Calculate Required Nominal Sample Size}
\usage{
calculate_cross_lagged_required_n(
  n_effective_needed,
  stability_coefficient,
  n_waves
)
}
\arguments{
\item{n_effective_needed}{Required effective sample size}

\item{stability_coefficient}{Stability coefficient}

\item{n_waves}{Number of waves}
}
\value{
Required nominal sample size
}
\description{
Calculate Required Nominal Sample Size
}
