% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_effects_power.R
\name{calculate_fixed_effects_design_effect}
\alias{calculate_fixed_effects_design_effect}
\title{Calculate Design Effect for Fixed Effects (REPORTING ONLY)}
\usage{
calculate_fixed_effects_design_effect(n_periods, icc)
}
\arguments{
\item{n_periods}{Number of time periods}

\item{icc}{Intraclass correlation coefficient}
}
\value{
Design effect multiplier
}
\description{
Calculate Design Effect for Fixed Effects (REPORTING ONLY)
}
