% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_effects_power.R
\name{calculate_fixed_effects_effective_n}
\alias{calculate_fixed_effects_effective_n}
\title{Calculate Effective Sample Size for Fixed Effects (REPORTING ONLY)}
\usage{
calculate_fixed_effects_effective_n(n_units, n_periods, icc)
}
\arguments{
\item{n_units}{Number of units}

\item{n_periods}{Number of periods}

\item{icc}{Intraclass correlation coefficient}
}
\value{
Effective sample size
}
\description{
Calculate Effective Sample Size for Fixed Effects (REPORTING ONLY)
}
