% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_effects_power.R
\name{calculate_fixed_effects_required_n}
\alias{calculate_fixed_effects_required_n}
\title{Calculate Required Number of Units}
\usage{
calculate_fixed_effects_required_n(n_effective_needed, n_periods, icc)
}
\arguments{
\item{n_effective_needed}{Required effective sample size}

\item{n_periods}{Number of periods}

\item{icc}{Intraclass correlation coefficient}
}
\value{
Required number of units
}
\description{
Calculate Required Number of Units
}
