% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_measures_power.R
\name{calculate_repeated_measures_efficiency}
\alias{calculate_repeated_measures_efficiency}
\title{Calculate Design Efficiency for Repeated Measures}
\usage{
calculate_repeated_measures_efficiency(
  n_timepoints,
  correlation_between_measures
)
}
\arguments{
\item{n_timepoints}{Number of timepoints}

\item{correlation_between_measures}{Correlation between measures}
}
\value{
Design efficiency multiplier
}
\description{
Calculate Design Efficiency for Repeated Measures
}
