% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_sem_power.R
\name{calculate_sem_complexity_factor}
\alias{calculate_sem_complexity_factor}
\title{Calculate SEM Complexity Factor}
\usage{
calculate_sem_complexity_factor(n_indicators_total, measurement_error)
}
\arguments{
\item{n_indicators_total}{Total number of indicators}

\item{measurement_error}{Measurement error proportion}
}
\value{
Complexity adjustment factor
}
\description{
Calculate SEM Complexity Factor
}
