% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_direct_effects_power.R
\name{calculate_sem_effective_sample_size}
\alias{calculate_sem_effective_sample_size}
\title{Calculate Effective Sample Size for SEM Complexity}
\usage{
calculate_sem_effective_sample_size(n, n_predictors, measurement_reliability)
}
\arguments{
\item{n}{Nominal sample size}

\item{n_predictors}{Number of predictors}

\item{measurement_reliability}{Average measurement reliability}
}
\value{
Effective sample size
}
\description{
Calculate Effective Sample Size for SEM Complexity
}
