% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohens_d_to_partial_r.R
\name{cohens_d_to_partial_r}
\alias{cohens_d_to_partial_r}
\title{Convert Cohen's d to Partial Correlation}
\usage{
cohens_d_to_partial_r(d, apply_discount = FALSE)
}
\arguments{
\item{d}{Cohen's d effect size}

\item{apply_discount}{Apply framework discount factor}
}
\value{
Partial correlation equivalent
}
\description{
Convert Cohen's d to Partial Correlation
}
