% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohens_f2_to_partial_r.R
\name{cohens_f2_to_partial_r}
\alias{cohens_f2_to_partial_r}
\title{Convert Cohen's f² to Partial Correlation}
\usage{
cohens_f2_to_partial_r(f2, apply_discount = FALSE)
}
\arguments{
\item{f2}{Cohen's f\00B2 effect size}

\item{apply_discount}{Apply framework discount factor}
}
\value{
Partial correlation equivalent
}
\description{
Convert Cohen's f² to Partial Correlation
}
