% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_power.R
\name{correlation_effect_size_calculation}
\alias{correlation_effect_size_calculation}
\title{Calculate Effect Size for Given Sample Size and Power}
\usage{
correlation_effect_size_calculation(n, power, alpha, two_tailed)
}
\arguments{
\item{n}{Sample size}

\item{power}{Target power}

\item{alpha}{Significance level}

\item{two_tailed}{Two-tailed test flag}
}
\value{
Required partial correlation
}
\description{
Calculate Effect Size for Given Sample Size and Power
}
