% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_power.R
\name{correlation_power_calculation}
\alias{correlation_power_calculation}
\title{Calculate Power for Given Correlation and Sample Size}
\usage{
correlation_power_calculation(r_partial, n, alpha, two_tailed)
}
\arguments{
\item{r_partial}{Partial correlation}

\item{n}{Sample size}

\item{alpha}{Significance level}

\item{two_tailed}{Two-tailed test flag}
}
\value{
Statistical power
}
\description{
Calculate Power for Given Correlation and Sample Size
}
