% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_power.R
\name{correlation_sample_size}
\alias{correlation_sample_size}
\title{Quick Correlation Sample Size Calculation}
\usage{
correlation_sample_size(r_partial, power = 0.8, alpha = 0.05)
}
\arguments{
\item{r_partial}{Partial correlation}

\item{power}{Target power (default = 0.8)}

\item{alpha}{Significance level (default = 0.05)}
}
\value{
Required sample size
}
\description{
Quick Correlation Sample Size Calculation
}
