% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_power.R
\name{correlation_sample_size_calculation}
\alias{correlation_sample_size_calculation}
\title{Calculate Sample Size for Given Correlation and Power}
\usage{
correlation_sample_size_calculation(r_partial, power, alpha, two_tailed)
}
\arguments{
\item{r_partial}{Partial correlation}

\item{power}{Target power}

\item{alpha}{Significance level}

\item{two_tailed}{Two-tailed test flag}
}
\value{
Required sample size
}
\description{
Calculate Sample Size for Given Correlation and Power
}
