% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_lagged_panel_power.R
\name{cross_lagged_panel_framework_power}
\alias{cross_lagged_panel_framework_power}
\title{Framework-Integrated Cross-Lagged Panel Power Analysis}
\usage{
cross_lagged_panel_framework_power(
  effect_size,
  effect_type = "r",
  n = NULL,
  power = 0.8,
  n_waves = 3,
  stability_coefficient = 0.6,
  alpha = 0.05,
  discount_factor = 0.75
)
}
\arguments{
\item{effect_size}{Effect size value}

\item{effect_type}{Type of effect size}

\item{n}{Sample size}

\item{power}{Target power}

\item{n_waves}{Number of waves}

\item{stability_coefficient}{Stability coefficient}

\item{alpha}{Significance level}

\item{discount_factor}{Framework discount factor}
}
\value{
Power analysis result
}
\description{
Framework-Integrated Cross-Lagged Panel Power Analysis
}
