% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_squared_to_partial_r.R
\name{eta_squared_to_partial_r}
\alias{eta_squared_to_partial_r}
\title{Convert Eta-squared to Partial Correlation}
\usage{
eta_squared_to_partial_r(eta_squared, apply_discount = FALSE)
}
\arguments{
\item{eta_squared}{Eta-squared effect size}

\item{apply_discount}{Apply framework discount factor}
}
\value{
Partial correlation equivalent
}
\description{
Convert Eta-squared to Partial Correlation
}
