% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_effects_power.R
\name{fixed_effects_effect_size_calculation}
\alias{fixed_effects_effect_size_calculation}
\title{Calculate Effect Size for Fixed Effects}
\usage{
fixed_effects_effect_size_calculation(n_effective, power, icc, alpha)
}
\arguments{
\item{n_effective}{Effective sample size}

\item{power}{Target power}

\item{icc}{Intraclass correlation coefficient}

\item{alpha}{Significance level}
}
\value{
Required within-person partial correlation
}
\description{
Calculate Effect Size for Fixed Effects
}
