% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_effects_power.R
\name{fixed_effects_framework_power}
\alias{fixed_effects_framework_power}
\title{Framework-Integrated Fixed Effects Power Analysis}
\usage{
fixed_effects_framework_power(
  effect_size,
  effect_type = "r",
  n_units = NULL,
  n_periods = 4,
  power = NULL,
  icc = 0.3,
  alpha = 0.05,
  discount_factor = 0.75
)
}
\arguments{
\item{effect_size}{Effect size value}

\item{effect_type}{Type of effect size}

\item{n_units}{Number of units}

\item{n_periods}{Number of periods}

\item{power}{Target power}

\item{icc}{Intraclass correlation coefficient}

\item{alpha}{Significance level}

\item{discount_factor}{Framework discount factor}
}
\value{
Power analysis result
}
\description{
Framework-Integrated Fixed Effects Power Analysis
}
