% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_effects_power.R
\name{fixed_effects_power_calculation}
\alias{fixed_effects_power_calculation}
\title{Calculate Power for Fixed Effects}
\usage{
fixed_effects_power_calculation(r_partial, n_effective, icc, alpha)
}
\arguments{
\item{r_partial}{Within-person partial correlation}

\item{n_effective}{Effective sample size}

\item{icc}{Intraclass correlation coefficient}

\item{alpha}{Significance level}
}
\value{
Statistical power
}
\description{
Calculate Power for Fixed Effects
}
