% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_effects_power.R
\name{fixed_effects_power_check}
\alias{fixed_effects_power_check}
\title{Quick Fixed Effects Power Calculation}
\usage{
fixed_effects_power_check(
  r_partial,
  n_units,
  n_periods = 4,
  icc = 0.3,
  alpha = 0.05
)
}
\arguments{
\item{r_partial}{Within-person partial correlation}

\item{n_units}{Number of units}

\item{n_periods}{Number of periods (default = 4)}

\item{icc}{Intraclass correlation coefficient (default = 0.3)}

\item{alpha}{Significance level (default = 0.05)}
}
\value{
Statistical power
}
\description{
Quick Fixed Effects Power Calculation
}
