% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_effects_power.R
\name{fixed_effects_sample_size}
\alias{fixed_effects_sample_size}
\title{Quick Fixed Effects Sample Size Calculation}
\usage{
fixed_effects_sample_size(
  r_partial,
  power = 0.8,
  n_periods = 4,
  icc = 0.3,
  alpha = 0.05
)
}
\arguments{
\item{r_partial}{Within-person partial correlation}

\item{power}{Target power (default = 0.8)}

\item{n_periods}{Number of periods (default = 4)}

\item{icc}{Intraclass correlation coefficient (default = 0.3)}

\item{alpha}{Significance level (default = 0.05)}
}
\value{
Required number of units
}
\description{
Quick Fixed Effects Sample Size Calculation
}
