% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_effects_power.R
\name{fixed_effects_sample_size_calculation}
\alias{fixed_effects_sample_size_calculation}
\title{Calculate Sample Size for Fixed Effects}
\usage{
fixed_effects_sample_size_calculation(r_partial, power, n_periods, icc, alpha)
}
\arguments{
\item{r_partial}{Within-person partial correlation}

\item{power}{Target power}

\item{n_periods}{Number of time periods}

\item{icc}{Intraclass correlation coefficient}

\item{alpha}{Significance level}
}
\value{
Required effective sample size
}
\description{
Calculate Sample Size for Fixed Effects
}
