% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework_effect_size.R
\name{framework_effect_size}
\alias{framework_effect_size}
\title{Framework-Integrated Effect Size Input}
\usage{
framework_effect_size(effect_size, effect_type = "r", apply_discount = TRUE)
}
\arguments{
\item{effect_size}{Input effect size}

\item{effect_type}{Type of effect size ("r", "d", "f2", "r_squared", "eta_squared")}

\item{apply_discount}{Whether to apply 0.75 discount factor}
}
\value{
Partial correlation with discount applied
}
\description{
Framework-Integrated Effect Size Input
}
