% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_effect_size.R
\name{interpret_effect_size}
\alias{interpret_effect_size}
\title{Effect Size Interpretation (Framework Standard)}
\usage{
interpret_effect_size(r, standard = "cohen")
}
\arguments{
\item{r}{Partial correlation}

\item{standard}{Interpretation standard ("cohen", "field")}
}
\value{
Interpretation string
}
\description{
Effect Size Interpretation (Framework Standard)
}
