% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_regression_power.R
\name{interpret_mediation_effect}
\alias{interpret_mediation_effect}
\title{Interpret Mediation Effect Size}
\usage{
interpret_mediation_effect(r_a, r_b)
}
\arguments{
\item{r_a}{A-path correlation}

\item{r_b}{B-path partial correlation}
}
\value{
Interpretation string
}
\description{
Interpret Mediation Effect Size
}
