% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_power.R
\name{linear_regression_effect_size_calculation}
\alias{linear_regression_effect_size_calculation}
\title{Calculate Effect Size for Linear Regression}
\usage{
linear_regression_effect_size_calculation(n, power, n_predictors, alpha)
}
\arguments{
\item{n}{Sample size}

\item{power}{Target power}

\item{n_predictors}{Number of predictors}

\item{alpha}{Significance level}
}
\value{
Required partial correlation
}
\description{
Calculate Effect Size for Linear Regression
}
