% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_power.R
\name{linear_regression_power_calculation}
\alias{linear_regression_power_calculation}
\title{Calculate Power for Linear Regression}
\usage{
linear_regression_power_calculation(r_partial, n, n_predictors, alpha)
}
\arguments{
\item{r_partial}{Partial correlation}

\item{n}{Sample size}

\item{n_predictors}{Number of predictors}

\item{alpha}{Significance level}
}
\value{
Statistical power
}
\description{
Calculate Power for Linear Regression
}
