% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression_power.R
\name{logistic_multi_predictor_sim_engine}
\alias{logistic_multi_predictor_sim_engine}
\title{Internal Simulation Engine for Multiple Logistic Regression (with correlated predictors)}
\usage{
logistic_multi_predictor_sim_engine(
  r_partial,
  n,
  n_predictors,
  alpha,
  n_sims,
  inter_predictor_cor = 0.3
)
}
\description{
Internal Simulation Engine for Multiple Logistic Regression (with correlated predictors)
}
\keyword{internal}
