% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression_power.R
\name{logistic_regression_sample_size_calculation}
\alias{logistic_regression_sample_size_calculation}
\title{Calculate Sample Size for Logistic Regression}
\usage{
logistic_regression_sample_size_calculation(
  r_partial,
  power,
  n_predictors,
  alpha
)
}
\arguments{
\item{r_partial}{Partial correlation}

\item{power}{Target power}

\item{n_predictors}{Number of predictors}

\item{alpha}{Significance level}
}
\value{
Required sample size
}
\description{
Calculate Sample Size for Logistic Regression
}
