% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_regression_power.R
\name{mediation_effect_size_calculation_a}
\alias{mediation_effect_size_calculation_a}
\title{Calculate Required A-Path Effect Size}
\usage{
mediation_effect_size_calculation_a(r_b, n, power, alpha, test_type)
}
\arguments{
\item{r_b}{B-path partial correlation}

\item{n}{Sample size}

\item{power}{Target power}

\item{alpha}{Significance level}

\item{test_type}{Type of mediation test}
}
\value{
Required a-path correlation
}
\description{
Calculate Required A-Path Effect Size
}
