% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_regression_power.R
\name{mediation_effect_size_calculation_b}
\alias{mediation_effect_size_calculation_b}
\title{Calculate Required B-Path Effect Size}
\usage{
mediation_effect_size_calculation_b(r_a, n, power, alpha, test_type)
}
\arguments{
\item{r_a}{A-path correlation}

\item{n}{Sample size}

\item{power}{Target power}

\item{alpha}{Significance level}

\item{test_type}{Type of mediation test}
}
\value{
Required b-path partial correlation
}
\description{
Calculate Required B-Path Effect Size
}
