% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_regression_power.R
\name{mediation_framework_power}
\alias{mediation_framework_power}
\title{Framework-Integrated Mediation Power Analysis}
\usage{
mediation_framework_power(
  effect_size_a,
  effect_size_b,
  effect_type = "r",
  n = NULL,
  power = 0.8,
  alpha = 0.05,
  discount_factor = 0.75
)
}
\arguments{
\item{effect_size_a}{A-path effect size value}

\item{effect_size_b}{B-path effect size value}

\item{effect_type}{Type of effect sizes}

\item{n}{Sample size}

\item{power}{Target power}

\item{alpha}{Significance level}

\item{discount_factor}{Framework discount factor}
}
\value{
Power analysis result
}
\description{
Framework-Integrated Mediation Power Analysis
}
