% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_regression_power.R
\name{mediation_power_calculation}
\alias{mediation_power_calculation}
\title{Calculate Power for Mediation Analysis}
\usage{
mediation_power_calculation(r_a, r_b, n, alpha, test_type)
}
\arguments{
\item{r_a}{A-path correlation}

\item{r_b}{B-path partial correlation}

\item{n}{Sample size}

\item{alpha}{Significance level}

\item{test_type}{Type of mediation test}
}
\value{
Statistical power
}
\description{
Calculate Power for Mediation Analysis
}
