% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_regression_power.R
\name{mediation_regression_power_check}
\alias{mediation_regression_power_check}
\title{Quick Mediation Power Calculation}
\usage{
mediation_regression_power_check(
  r_a,
  r_b,
  n,
  alpha = 0.05,
  discount_factor = 0.75
)
}
\arguments{
\item{r_a}{A-path correlation}

\item{r_b}{B-path partial correlation}

\item{n}{Sample size}

\item{alpha}{Significance level (default = 0.05)}

\item{discount_factor}{Framework discount factor}
}
\value{
Statistical power
}
\description{
Quick Mediation Power Calculation
}
