% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_regression_power.R
\name{mediation_sample_size_calculation}
\alias{mediation_sample_size_calculation}
\title{Calculate Sample Size for Mediation Analysis}
\usage{
mediation_sample_size_calculation(r_a, r_b, power, alpha, test_type)
}
\arguments{
\item{r_a}{A-path correlation}

\item{r_b}{B-path partial correlation}

\item{power}{Target power}

\item{alpha}{Significance level}

\item{test_type}{Type of mediation test}
}
\value{
Required sample size
}
\description{
Calculate Sample Size for Mediation Analysis
}
