% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_sem_power.R
\name{mediation_sem_framework_power}
\alias{mediation_sem_framework_power}
\title{Framework-Integrated SEM Mediation Power Analysis}
\usage{
mediation_sem_framework_power(
  effect_size_a,
  effect_size_b,
  effect_type = "r",
  n = NULL,
  power = 0.8,
  measurement_error = 0.1,
  n_indicators_total = 9,
  alpha = 0.05,
  discount_factor = 0.75
)
}
\arguments{
\item{effect_size_a}{Effect size for path a}

\item{effect_size_b}{Effect size for path b}

\item{effect_type}{Type of effect size}

\item{n}{Sample size}

\item{power}{Target power}

\item{measurement_error}{Measurement error proportion}

\item{n_indicators_total}{Total indicators}

\item{alpha}{Significance level}

\item{discount_factor}{Framework discount factor}
}
\value{
Power analysis result
}
\description{
Framework-Integrated SEM Mediation Power Analysis
}
