% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_sem_power.R
\name{mediation_sem_power_calculation}
\alias{mediation_sem_power_calculation}
\title{Calculate Power for SEM Mediation}
\usage{
mediation_sem_power_calculation(r_a, r_b, n, sem_complexity_factor, alpha)
}
\arguments{
\item{r_a}{Path a coefficient}

\item{r_b}{Path b coefficient}

\item{n}{Sample size}

\item{sem_complexity_factor}{SEM complexity adjustment}

\item{alpha}{Significance level}
}
\value{
Statistical power
}
\description{
Calculate Power for SEM Mediation
}
