% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_sem_power.R
\name{mediation_sem_power_check}
\alias{mediation_sem_power_check}
\title{Quick Mediation SEM Power Check}
\usage{
mediation_sem_power_check(
  r_a,
  r_b,
  n,
  alpha = 0.05,
  discount_factor = 0.75,
  measurement_reliability = 0.9
)
}
\arguments{
\item{r_a}{Latent path a (X->M) coefficient}

\item{r_b}{Latent path b (M->Y|X) coefficient}

\item{n}{Sample size}

\item{alpha}{Significance level}

\item{discount_factor}{Conservative discount factor}

\item{measurement_reliability}{Average reliability of constructs}
}
\value{
Power
}
\description{
Quick Mediation SEM Power Check
}
