% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_models_power.R
\name{mixed_model_level1_sim_engine}
\alias{mixed_model_level1_sim_engine}
\title{Internal Simulation Engine for Level-1 Effects with Correlated Predictors}
\usage{
mixed_model_level1_sim_engine(
  r_partial,
  n_groups,
  n_per_group,
  icc,
  alpha,
  n_sims,
  inter_predictor_cor = 0
)
}
\description{
Internal Simulation Engine for Level-1 Effects with Correlated Predictors
}
\keyword{internal}
