% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_models_power.R
\name{mixed_models_framework_power}
\alias{mixed_models_framework_power}
\title{Framework-Integrated Mixed Models Power Analysis}
\usage{
mixed_models_framework_power(
  effect_size,
  effect_type = "r",
  n_groups = NULL,
  power = 0.8,
  n_per_group = 10,
  icc = 0.05,
  test_level = "level1",
  alpha = 0.05,
  discount_factor = 0.75
)
}
\arguments{
\item{effect_size}{Effect size value}

\item{effect_type}{Type of effect size}

\item{n_groups}{Number of groups}

\item{power}{Target power}

\item{n_per_group}{Observations per group}

\item{icc}{Intraclass correlation}

\item{test_level}{Level being tested}

\item{alpha}{Significance level}

\item{discount_factor}{Framework discount factor}
}
\value{
Power analysis result
}
\description{
Framework-Integrated Mixed Models Power Analysis
}
