% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_models_power.R
\name{mixed_models_power}
\alias{mixed_models_power}
\title{Mixed Models Power Analysis}
\usage{
mixed_models_power(
  r_partial = NULL,
  n_groups = NULL,
  n_per_group = 10,
  power = NULL,
  icc = 0.05,
  alpha = 0.05,
  discount_factor = 0.75,
  test_level = "level1",
  n_sims_l1 = 500,
  effect_input = NULL,
  effect_type = "r",
  inter_predictor_cor = 0
)
}
\arguments{
\item{r_partial}{Partial correlation for the effect of interest.}

\item{n_groups}{Number of groups (Level-2 units).}

\item{n_per_group}{Average number of observations per group (Level-1 units).}

\item{power}{Target statistical power.}

\item{icc}{Intraclass Correlation Coefficient.}

\item{alpha}{Significance level.}

\item{discount_factor}{Conservative planning discount factor.}

\item{test_level}{The level of the predictor: "level1" or "level2".}

\item{n_sims_l1}{Number of simulations for Level-1 power estimation.}

\item{effect_input}{Alternative raw effect size.}

\item{effect_type}{Type of \code{effect_input}.}

\item{inter_predictor_cor}{The estimated correlation between predictors (default = 0).}
}
\value{
A list with power analysis results.
}
\description{
Performs power analysis for mixed-effects (multilevel) models.
}
