% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_models_power.R
\name{mixed_models_power_check}
\alias{mixed_models_power_check}
\title{Quick Mixed Models Power Check}
\usage{
mixed_models_power_check(
  r_partial,
  n_groups,
  alpha = 0.05,
  n_per_group = 10,
  icc = 0.05,
  test_level = "level1"
)
}
\arguments{
\item{r_partial}{Partial correlation for the predictor.}

\item{n_groups}{Number of groups.}

\item{alpha}{Significance level (default = 0.05).}

\item{n_per_group}{Number of observations per group (default = 10).}

\item{icc}{Intraclass correlation (default = 0.05).}

\item{test_level}{The level of the effect, "level1" or "level2".}
}
\value{
Achieved statistical power.
}
\description{
Convenience function to calculate power.
}
