% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio_to_partial_r.R
\name{odds_ratio_to_partial_r}
\alias{odds_ratio_to_partial_r}
\title{Convert Odds Ratio to Partial Correlation}
\usage{
odds_ratio_to_partial_r(or, apply_discount = FALSE)
}
\arguments{
\item{or}{The odds ratio from a logistic regression.}

\item{apply_discount}{Apply framework discount factor (default: FALSE).}
}
\value{
Partial correlation equivalent (r).
}
\description{
Convert Odds Ratio to Partial Correlation
}
