% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_r_to_cohens_d.R
\name{partial_r_to_cohens_d}
\alias{partial_r_to_cohens_d}
\title{Convert Partial Correlation to Cohen's d}
\usage{
partial_r_to_cohens_d(r)
}
\arguments{
\item{r}{Partial correlation}
}
\value{
Cohen's d equivalent
}
\description{
Convert Partial Correlation to Cohen's d
}
