% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_r_to_cohens_f2.R
\name{partial_r_to_cohens_f2}
\alias{partial_r_to_cohens_f2}
\title{Convert Partial Correlation to Cohen's f².}
\usage{
partial_r_to_cohens_f2(r)
}
\arguments{
\item{r}{Partial correlation}
}
\value{
Cohen's f\00B2 equivalent
}
\description{
Convert Partial Correlation to Cohen's f².
}
