% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_lagged_panel_power.R
\name{print.cross_lagged_panel_power_analysis}
\alias{print.cross_lagged_panel_power_analysis}
\title{Print Method for Cross-Lagged Panel Power Analysis}
\usage{
\method{print}{cross_lagged_panel_power_analysis}(x, ...)
}
\arguments{
\item{x}{Cross-lagged panel power analysis result}

\item{...}{Additional arguments}
}
\description{
Print Method for Cross-Lagged Panel Power Analysis
}
