% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_regression_power.R
\name{print.mediation_regression_power_analysis}
\alias{print.mediation_regression_power_analysis}
\title{Print Method for Mediation Regression Power Analysis}
\usage{
\method{print}{mediation_regression_power_analysis}(x, ...)
}
\arguments{
\item{x}{Mediation regression power analysis result}

\item{...}{Additional arguments}
}
\description{
Print Method for Mediation Regression Power Analysis
}
