% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_models_power.R
\name{print.mixed_models_power_analysis}
\alias{print.mixed_models_power_analysis}
\title{Print Method for Mixed Models Power Analysis}
\usage{
\method{print}{mixed_models_power_analysis}(x, ...)
}
\arguments{
\item{x}{Mixed models power analysis result}

\item{...}{Additional arguments}
}
\description{
Print Method for Mixed Models Power Analysis
}
