% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_direct_effects_power.R
\name{print.sem_direct_effects_power_analysis}
\alias{print.sem_direct_effects_power_analysis}
\title{Print Method for SEM Direct Effects Power Analysis}
\usage{
\method{print}{sem_direct_effects_power_analysis}(x, ...)
}
\arguments{
\item{x}{SEM direct effects power analysis result}

\item{...}{Additional arguments}
}
\description{
Print Method for SEM Direct Effects Power Analysis
}
