% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_measures_power.R
\name{repeated_measures_effect_size_calculation}
\alias{repeated_measures_effect_size_calculation}
\title{Calculate Effect Size for Repeated Measures}
\usage{
repeated_measures_effect_size_calculation(
  n,
  power,
  n_timepoints,
  correlation_between_measures,
  alpha
)
}
\arguments{
\item{n}{Sample size}

\item{power}{Target power}

\item{n_timepoints}{Number of timepoints}

\item{correlation_between_measures}{Correlation between measures}

\item{alpha}{Significance level}
}
\value{
Required partial correlation
}
\description{
Calculate Effect Size for Repeated Measures
}
