% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_measures_power.R
\name{repeated_measures_framework_power}
\alias{repeated_measures_framework_power}
\title{Framework-Integrated Repeated Measures Power Analysis}
\usage{
repeated_measures_framework_power(
  effect_size,
  effect_type = "r",
  n = NULL,
  power = 0.8,
  n_timepoints = 2,
  correlation_between_measures = 0.5,
  alpha = 0.05,
  discount_factor = 0.75
)
}
\arguments{
\item{effect_size}{Effect size value}

\item{effect_type}{Type of effect size}

\item{n}{Sample size}

\item{power}{Target power}

\item{n_timepoints}{Number of timepoints}

\item{correlation_between_measures}{Correlation between measures}

\item{alpha}{Significance level}

\item{discount_factor}{Framework discount factor}
}
\value{
Power analysis result
}
\description{
Framework-Integrated Repeated Measures Power Analysis
}
