% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_measures_power.R
\name{repeated_measures_power_check}
\alias{repeated_measures_power_check}
\title{Quick Repeated Measures Power Calculation}
\usage{
repeated_measures_power_check(
  r_partial,
  n,
  n_timepoints = 2,
  correlation_between_measures = 0.5,
  alpha = 0.05
)
}
\arguments{
\item{r_partial}{Partial correlation}

\item{n}{Sample size}

\item{n_timepoints}{Number of timepoints (default = 2)}

\item{correlation_between_measures}{Correlation between measures (default = 0.5)}

\item{alpha}{Significance level (default = 0.05)}
}
\value{
Statistical power
}
\description{
Quick Repeated Measures Power Calculation
}
