% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_measures_power.R
\name{repeated_measures_sample_size}
\alias{repeated_measures_sample_size}
\title{Quick Repeated Measures Sample Size Calculation}
\usage{
repeated_measures_sample_size(
  r_partial,
  power = 0.8,
  n_timepoints = 2,
  correlation_between_measures = 0.5,
  alpha = 0.05
)
}
\arguments{
\item{r_partial}{Partial correlation}

\item{power}{Target power (default = 0.8)}

\item{n_timepoints}{Number of timepoints (default = 2)}

\item{correlation_between_measures}{Correlation between measures (default = 0.5)}

\item{alpha}{Significance level (default = 0.05)}
}
\value{
Required sample size
}
\description{
Quick Repeated Measures Sample Size Calculation
}
