% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_measures_power.R
\name{repeated_measures_sample_size_calculation}
\alias{repeated_measures_sample_size_calculation}
\title{Calculate Sample Size for Repeated Measures}
\usage{
repeated_measures_sample_size_calculation(
  r_partial,
  power,
  alpha,
  correlation_between_measures,
  n_timepoints
)
}
\arguments{
\item{r_partial}{Partial correlation coefficient, analogous to sqrt(partial eta-squared) (NULL to calculate)}

\item{power}{Statistical power (NULL to calculate, default = 0.8)}

\item{alpha}{Significance level (default = 0.05)}

\item{correlation_between_measures}{Average correlation between measures (default = 0.5).}

\item{n_timepoints}{Number of repeated measures (default = 2). Note: engine uses paired t-test, best for 2 timepoints.}
}
\value{
Sample Size
}
\description{
Calculate Sample Size for Repeated Measures
}
