% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_direct_effects_power.R
\name{sem_direct_effects_effect_size_calculation}
\alias{sem_direct_effects_effect_size_calculation}
\title{Calculate Effect Size for SEM Direct Effects (Approximation)}
\usage{
sem_direct_effects_effect_size_calculation(
  n,
  power,
  n_predictors,
  alpha,
  measurement_reliability
)
}
\arguments{
\item{n}{Sample size}

\item{power}{Target power}

\item{n_predictors}{Number of predictors}

\item{alpha}{Significance level}

\item{measurement_reliability}{Average measurement reliability}
}
\value{
Required partial correlation
}
\description{
Calculate Effect Size for SEM Direct Effects (Approximation)
}
