% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_direct_effects_power.R
\name{sem_direct_effects_framework_power}
\alias{sem_direct_effects_framework_power}
\title{Framework-Integrated SEM Direct Effects Power Analysis}
\usage{
sem_direct_effects_framework_power(
  effect_size,
  effect_type = "r",
  n = NULL,
  power = 0.8,
  n_predictors = 3,
  alpha = 0.05,
  discount_factor = 0.75
)
}
\arguments{
\item{effect_size}{Effect size value}

\item{effect_type}{Type of effect size}

\item{n}{Sample size}

\item{power}{Target power}

\item{n_predictors}{Number of predictors}

\item{alpha}{Significance level}

\item{discount_factor}{Framework discount factor}
}
\value{
Power analysis result
}
\description{
Framework-Integrated SEM Direct Effects Power Analysis
}
