% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_sem_power.R
\name{sem_mediation_sample_size_calculation}
\alias{sem_mediation_sample_size_calculation}
\title{SEM Mediation Sample Size Calculation
This internal function is used by the main wrapper to solve for sample size.}
\usage{
sem_mediation_sample_size_calculation(r_a, r_b, power, alpha, reliability)
}
\arguments{
\item{r_a}{X -> Mediator path coefficient (NULL to calculate)}

\item{r_b}{Mediator -> Y path coefficient (controlling for X)}

\item{power}{Statistical power}

\item{alpha}{Significance level (default = 0.05)}

\item{reliability}{Average reliability of constructs}
}
\value{
Sample size
}
\description{
SEM Mediation Sample Size Calculation
This internal function is used by the main wrapper to solve for sample size.
}
