% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unified_effect_size_table.R
\name{unified_effect_size_table}
\alias{unified_effect_size_table}
\title{Unified Effect Size Table for Framework}
\usage{
unified_effect_size_table(
  effect_values,
  input_type = "r",
  apply_discount = TRUE
)
}
\arguments{
\item{effect_values}{Vector of effect sizes}

\item{input_type}{Type of input effect sizes}

\item{apply_discount}{Apply framework discount factor}
}
\value{
Data frame with unified conversions
}
\description{
Unified Effect Size Table for Framework
}
