% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon_signed_rank_power.R
\name{wilcoxon_power_calculation}
\alias{wilcoxon_power_calculation}
\title{Calculate Power for Wilcoxon Signed-Rank Test}
\usage{
wilcoxon_power_calculation(r_partial, n, alpha, two_tailed, asymptotic)
}
\arguments{
\item{r_partial}{Partial correlation}

\item{n}{Sample size}

\item{alpha}{Significance level}

\item{two_tailed}{Two-tailed test flag}

\item{asymptotic}{Use asymptotic approximation}
}
\value{
Statistical power
}
\description{
Calculate Power for Wilcoxon Signed-Rank Test
}
