% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon_signed_rank_power.R
\name{wilcoxon_power_check}
\alias{wilcoxon_power_check}
\title{Quick Wilcoxon Power Calculation}
\usage{
wilcoxon_power_check(r_partial, n, alpha = 0.05)
}
\arguments{
\item{r_partial}{Partial correlation}

\item{n}{Sample size}

\item{alpha}{Significance level (default = 0.05)}
}
\value{
Statistical power
}
\description{
Quick Wilcoxon Power Calculation
}
