% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon_signed_rank_power.R
\name{wilcoxon_signed_rank_framework_power}
\alias{wilcoxon_signed_rank_framework_power}
\title{Framework-Integrated Wilcoxon Power Analysis}
\usage{
wilcoxon_signed_rank_framework_power(
  effect_size,
  effect_type = "r",
  n = NULL,
  power = NULL,
  alpha = 0.05,
  discount_factor = 0.75
)
}
\arguments{
\item{effect_size}{Effect size value}

\item{effect_type}{Type of effect size}

\item{n}{Sample size}

\item{power}{Target power}

\item{alpha}{Significance level}

\item{discount_factor}{Framework discount factor}
}
\value{
Power analysis result
}
\description{
Framework-Integrated Wilcoxon Power Analysis
}
