% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon_signed_rank_power.R
\name{wilcoxon_specific_conversions}
\alias{wilcoxon_specific_conversions}
\title{Wilcoxon-Specific Effect Size Conversions}
\usage{
wilcoxon_specific_conversions(r_partial)
}
\arguments{
\item{r_partial}{Partial correlation}
}
\value{
Named list of Wilcoxon-specific metrics
}
\description{
Wilcoxon-Specific Effect Size Conversions
}
