% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_regression_power.R,
%   R/mediation_sem_power.R
\name{interpret_mediation_effect}
\alias{interpret_mediation_effect}
\title{Interpret Mediation Effect Size}
\usage{
interpret_mediation_effect(r_a, r_b)

interpret_mediation_effect(r_a, r_b)
}
\arguments{
\item{r_a}{Latent path a (X->M) coefficient}

\item{r_b}{Latent path b (M->Y|X) coefficient}
}
\value{
Interpretation string

Mediation Effect
}
\description{
Interpret Mediation Effect Size

Interpret Mediation Effect
}
