% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression_power.R
\name{logistic_regression_sample_size}
\alias{logistic_regression_sample_size}
\title{Quick Logistic Regression Sample Size Calculation}
\usage{
logistic_regression_sample_size(
  r_partial = NULL,
  power = 0.8,
  n_predictors = 1,
  ...
)
}
\arguments{
\item{r_partial}{Partial correlation for the predictor of interest.}

\item{power}{Target statistical power (defaults to 0.8).}

\item{n_predictors}{Total number of predictors in the model.}

\item{...}{Other arguments passed to \code{logistic_regression_power}, such as \code{effect_input}, \code{effect_type}, \code{alpha}, or \code{inter_predictor_cor}.}
}
\value{
Required sample size (N).
}
\description{
Quick Logistic Regression Sample Size Calculation
}
