% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_sem_power.R
\name{mediation_sem_sample_size}
\alias{mediation_sem_sample_size}
\title{Quick Mediation SEM Sample Size Calculation}
\usage{
mediation_sem_sample_size(
  r_a,
  r_b,
  power = 0.8,
  alpha = 0.05,
  discount_factor = 0.75,
  measurement_reliability = 0.9
)
}
\arguments{
\item{r_a}{Latent path a (X->M) coefficient}

\item{r_b}{Latent path b (M->Y|X) coefficient}

\item{power}{Statistical power}

\item{alpha}{Significance level}

\item{discount_factor}{Conservative discount factor}

\item{measurement_reliability}{Average reliability of constructs}
}
\value{
Sample size
}
\description{
Quick Mediation SEM Sample Size Calculation
}
