% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_models_power.R
\name{mixed_models_effect_size_calculation}
\alias{mixed_models_effect_size_calculation}
\title{Calculate Effect Size for Mixed Models}
\usage{
mixed_models_effect_size_calculation(
  n_groups,
  power,
  n_per_group,
  icc,
  test_level,
  alpha
)
}
\arguments{
\item{n_groups}{Number of groups}

\item{power}{Target power}

\item{n_per_group}{Observations per group}

\item{icc}{Intraclass correlation}

\item{test_level}{Level being tested}

\item{alpha}{Significance level}
}
\value{
Required partial correlation
}
\description{
Calculate Effect Size for Mixed Models
}
