% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_models_power.R
\name{mixed_models_power_calculation}
\alias{mixed_models_power_calculation}
\title{Calculate Power for Mixed Models}
\usage{
mixed_models_power_calculation(
  r_partial,
  n_groups,
  n_per_group,
  icc,
  alpha,
  test_level
)
}
\arguments{
\item{r_partial}{Partial correlation}

\item{n_groups}{Number of groups}

\item{n_per_group}{Observations per group}

\item{icc}{Intraclass correlation}

\item{alpha}{Significance level}

\item{test_level}{Level being tested}
}
\value{
Statistical power
}
\description{
Calculate Power for Mixed Models
}
