% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_models_power.R
\name{mixed_models_power_check}
\alias{mixed_models_power_check}
\title{Quick Mixed Models Power Calculation}
\usage{
mixed_models_power_check(
  r_partial,
  n_groups,
  n_per_group = 10,
  icc = 0.05,
  test_level = "level1"
)
}
\arguments{
\item{r_partial}{Partial correlation}

\item{n_groups}{Number of groups}

\item{n_per_group}{Observations per group (default = 10)}

\item{icc}{Intraclass correlation (default = 0.05)}

\item{test_level}{Level being tested (default = "level1")}
}
\value{
Statistical power
}
\description{
Quick Mixed Models Power Calculation
}
