% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_models_power.R
\name{mixed_models_sample_size_calculation}
\alias{mixed_models_sample_size_calculation}
\title{Calculate Number of Groups for Mixed Models}
\usage{
mixed_models_sample_size_calculation(
  r_partial,
  power,
  n_per_group,
  icc,
  test_level,
  alpha
)
}
\arguments{
\item{r_partial}{Partial correlation}

\item{power}{Target power}

\item{n_per_group}{Observations per group}

\item{icc}{Intraclass correlation}

\item{test_level}{Level being tested}

\item{alpha}{Significance level}
}
\value{
Required number of groups
}
\description{
Calculate Number of Groups for Mixed Models
}
