% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_models_power.R
\name{print.mixed_models_power_analysis}
\alias{print.mixed_models_power_analysis}
\title{Print method for mixed models power analysis}
\usage{
\method{print}{mixed_models_power_analysis}(x, ...)
}
\arguments{
\item{x}{A \code{mixed_models_power_analysis} object.}

\item{...}{Additional arguments passed to \code{print}.}
}
\description{
Print method for mixed models power analysis
}
