% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_measures_power.R
\name{repeated_measures_power_calculation}
\alias{repeated_measures_power_calculation}
\title{Calculate Power for Repeated Measures}
\usage{
repeated_measures_power_calculation(
  r_partial,
  n,
  n_timepoints,
  correlation_between_measures,
  alpha
)
}
\arguments{
\item{r_partial}{Partial correlation}

\item{n}{Sample size}

\item{n_timepoints}{Number of timepoints}

\item{correlation_between_measures}{Correlation between measures}

\item{alpha}{Significance level}
}
\value{
Statistical power
}
\description{
Calculate Power for Repeated Measures
}
