% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_measures_power.R
\name{repeated_measures_sample_size_calculation}
\alias{repeated_measures_sample_size_calculation}
\title{Calculate Sample Size for Repeated Measures}
\usage{
repeated_measures_sample_size_calculation(
  r_partial,
  power,
  n_timepoints,
  correlation_between_measures,
  alpha
)
}
\arguments{
\item{r_partial}{Partial correlation}

\item{power}{Target power}

\item{n_timepoints}{Number of timepoints}

\item{correlation_between_measures}{Correlation between measures}

\item{alpha}{Significance level}
}
\value{
Required sample size
}
\description{
Calculate Sample Size for Repeated Measures
}
